+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2020-07-02
-
+## [0.2.0] - 2020-07-03
 ### Added
-* making full screen screenshot.
+* making region screenshot.
+* making screenshot of first screen.
+* making screenshot of second screen.
+* copy screenshot to clipboard.
+
+## [0.1.0] - 2020-07-02
+### Added
+* making screenshot of all screens.